fix: 이벤트 DTO 필드 추가 및 null 체크 방어 로직 보완

+++ services/notification-service/src/main/java/com/bankrupang/sanjijk/notification/infrastructure/messaging/dto/BidOvertakenEvent.java
@@ -12,7 +12,7 @@ public class BidOvertakenEvent {
     private String auctionId;
     private String auctionTitle;
     private String previousBidderId;
-    private Long newPrice;
-    private Long nextMinPrice;
+    private Integer newPrice;
+    private Integer nextMinPrice;
     private LocalDateTime occurredAt;
 }
